third_party/WebKit/LayoutTests/fast/dom/CSSStyleDeclaration/css-properties-case-sensitive-expected.txt

This test checks that access to the CSS float property via JavaScript properties on DOM elements is case sensitive.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

normal cases


PASS element.style.float is 'left'
PASS element.style.Float is undefined.


"css" prefix


PASS element.style.cssFloat is 'left'
PASS element.style.CssFloat is undefined.
PASS element.style.Cssfloat is undefined.
PASS element.style.cssfloat is undefined.
PASS successfullyParsed is true

TEST COMPLETE
